\name{print.nonsigletters}
\alias{print.nonsigletters}

\title{ Print function for class "nonsigletters" }
\description{
Print function for nonsigletters.
}

\usage{
\method{print}{nonsigletters}(x , \dots )
}

\arguments{
  \item{x}{ an object of class "nonsigletters", the output of function \code{nonsigletters} }
  \item{\dots}{further arguments to be passed to \code{print}}
}

\author{ Frank Schaarschmidt }

}
\keyword{ htest }