\name{print.pairwiseCI}
\alias{print.pairwiseCI}

\title{ Print function for "pairwiseCI" }
\description{
Print out confidence intervals calculated using pairwiseCI.
}

\usage{
\method{print}{pairwiseCI}(x , digits=4, \dots )
}

\arguments{
  \item{x}{ an object of class "pairwiseCI", the output of function pairwiseCI() }
  \item{digits}{ integer, to which decimal output shall be rounded }
  \item{\dots}{further arguments to be passed to \code{print}}
}

\author{ Frank Schaarschmidt }

}
\keyword{ print }

