\name{print.pairwiseTest}
\alias{print.pairwiseTest}
\title{ Print function for "pairwiseTest" }
\description{
Print function for pairwiseTest objects
 }
\usage{
\method{print}{pairwiseTest}(x, digits = 4, \dots)
}
\arguments{
  \item{x}{ an object of class "pairwiseTest" }
  \item{digits}{ digits for rounding }
  \item{\dots}{further arguments to be passed to \code{print}}
}

\author{ Frank Schaarschmidt }
\keyword{print}

