\name{print.summary.pairwiseCI}
\alias{print.summary.pairwiseCI}

\title{ Print function for "summary.pairwiseCI" }
\description{
 Print function for summary.pairwiseCI 
}
\usage{
\method{print}{summary.pairwiseCI}(x, \dots)
}
\arguments{
  \item{x}{ an object of class "summary.pairwiseCI", created by the function \code{summary.pairwiseCI} }
  \item{\dots}{ further arguments to be passed to \code{print} }
}


\keyword{print}