\name{print.summary.pairwiseTest}
\alias{print.summary.pairwiseTest}

\title{ Print function for "summary.pairwiseTest" }
\description{
 Print function for summary.pairwiseCI 
}
\usage{
\method{print}{summary.pairwiseTest}(x, \dots)
}
\arguments{
  \item{x}{ an object of class "summary.pairwiseTest", created by the function \code{summary.pairwiseTest} }
  \item{\dots}{ further arguments to be passed to \code{print} }
}


\keyword{print}